requirements.txt
# Устанавливаем актуальную ветку aiogram 3.x для Telegram-бота.
aiogram==3.26.0

# Устанавливаем библиотеку для чтения переменных из файла .env.
python-dotenv>=1.0,<2.0

# Явно добавляем aiohttp, потому что через него будем запрашивать погодный API.
aiohttp>=3.9,<4.0
